move _HashStack back to hashing.py
