+++ lib/streamlit/caching/cache_errors.py
@@ -12,13 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import collections
 import enum
 import inspect
 import types
-from typing import Any, List, Callable, Optional
+from typing import Any, List, Optional
 
-from streamlit import type_util, util
+from streamlit import type_util
 from streamlit.errors import (
     StreamlitAPIWarning,
     StreamlitAPIException,
@@ -33,51 +32,6 @@ class CacheType(enum.Enum):
 class HashReason(enum.Enum):
     CACHING_FUNC_ARGS = "caching_func_args"
     CACHING_FUNC_BODY = "caching_func_body"
-
-
-class HashStack:
-    """Stack of what has been hashed, for debug and circular reference detection.
-
-    This internally keeps 1 stack per thread.
-
-    Internally, this stores the ID of pushed objects rather than the objects
-    themselves because otherwise the "in" operator inside __contains__ would
-    fail for objects that don't return a boolean for "==" operator. For
-    example, arr == 10 where arr is a NumPy array returns another NumPy array.
-    This causes the "in" to crash since it expects a boolean.
-    """
-
-    def __init__(self):
-        self._stack: collections.OrderedDict[int, List[Any]] = collections.OrderedDict()
-
-        # The reason why we're doing this hashing, for debug purposes.
-        self.hash_reason: Optional[HashReason] = None
-
-        # Either a function or a code block, depending on whether the reason is
-        # due to hashing part of a function (i.e. body, args, output) or an
-        # st.Cache codeblock.
-        self.hash_source: Optional[Callable[..., Any]] = None
-
-    def __repr__(self) -> str:
-        return util.repr_(self)
-
-    def push(self, val: Any):
-        self._stack[id(val)] = val
-
-    def pop(self):
-        self._stack.popitem()
-
-    def __contains__(self, val: Any):
-        return id(val) in self._stack
-
-    def pretty_print(self):
-        def to_str(v):
-            try:
-                return "Object of type %s: %s" % (type_util.get_fqn_type(v), str(v))
-            except:
-                return "<Unable to convert item to string>"
-
-        return "\n".join(to_str(x) for x in reversed(self._stack.values()))
 
 
 class UnhashableTypeError(Exception):
